Get these tests finally going!
Use nunit for the tracking collection tests, as it
handles things better there. Update to xunit 2.1 while
we're at it, and kill the old tracking collection tests,
since the new ones have full coverage of the collection.
