+++ src/TrackingCollectionTests/ITestOutputHelper.cs
@@ -1,0 +1,5 @@
+public interface ITestOutputHelper
+{
+    void WriteLine(string message);
+    void WriteLine(string format, params object[] args);
+}
