+++ src/TrackingCollectionTests/packages.config
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NUnit" version="2.6.4" targetFramework="net452" />
+  <package id="NUnit.Runners" version="2.6.4" targetFramework="net452" />
+  <package id="NUnitTestAdapter" version="2.0.0" targetFramework="net452" />
+  <package id="Rx-Core" version="2.2.5-custom" targetFramework="net452" />
+  <package id="Rx-Interfaces" version="2.2.5-custom" targetFramework="net452" />
+  <package id="Rx-Linq" version="2.2.5-custom" targetFramework="net452" />
+  <package id="Rx-Main" version="2.2.5-custom" targetFramework="net452" />
+  <package id="Rx-PlatformServices" version="2.2.5-custom" targetFramework="net452" />
+  <package id="Rx-XAML" version="2.2.5-custom" targetFramework="net452" />
+</packages>
